requirements.txt
asgiref==3.8.1
Django==5.0.5
django-colorfield==0.11.0
django-filter==24.2
django-jazzmin==3.0.0
django-otp==1.5.0
djangorestframework==3.15.1
django-summernote==0.8.20.0
Markdown==3.6
pillow==10.3.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.0
typing_extensions==4.11.0
